content/base/test/csp/test_csp_report.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=548193
-->
<head>
  <title>Test for Bug 548193</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<iframe style="width:200px;height:200px;" id='cspframe'></iframe>
<script class="testbody" type="text/javascript">

/*
 * Description of the test:
 * We try to load an inline-src using a policy that constrains
 * all scripts from running (script-src 'none'). We verify that
 * the generated csp-report contains the expceted values. If any
 * of the JSON is not formatted properly (e.g. not properly escaped)
 * then JSON.parse will fail, which allows to pinpoint such errors
 * in the catch block, and the test will fail. Since we use an
 * observer, we can set the actual report-uri to a foo value.
 */

const testfile = "tests/content/base/test/csp/file_csp_report.html";
const reportURI = "http://mochi.test:8888/foo.sjs";
const policy = "script-src 'none'; report-uri " + reportURI;
const docUri = "http://mochi.test:8888/tests/content/base/test/csp/file_csp_testserver.sjs" +
               "?file=tests/content/base/test/csp/file_csp_report.html" +
               "&csp=script-src%20%27none%27%3B%20report-uri%20http%3A//mochi.test%3A8888/foo.sjs";

window.checkResults = function(reportObj) {
  var cspReport = reportObj["csp-report"];

  // The following uris' fragments should be stripped before reporting:
  //    * document-uri
  //    * blocked-uri
  //    * source-file
  // see http://www.w3.org/TR/CSP11/#violation-reports
  is(cspReport["document-uri"], docUri, "Incorrect document-uri");

  // we can not test for the whole referrer since it includes platform specific information
  ok(cspReport["referrer"].startsWith("http://mochi.test:8888/tests/content/base/test/csp/test_csp_report.html"),
     "Incorrect referrer");

  is(cspReport["blocked-uri"], "self", "Incorrect blocked-uri");

  is(cspReport["violated-directive"], "script-src 'none'", "Incorrect violated-directive");

  is(cspReport["original-policy"], "script-src 'none'; report-uri http://mochi.test:8888/foo.sjs",
     "Incorrect original-policy");

  is(cspReport["source-file"], docUri, "Incorrect source-file");

  is(cspReport["script-sample"], "\n    var foo = \"propEscFoo\";\n    var bar...",
     "Incorrect script-sample");

  is(cspReport["line-number"], "7", "Incorrect line-number");
}

// This is used to watch requests go out so we can see if the report is
// sent correctly
function examiner() {
  SpecialPowers.addObserver(this, "http-on-opening-request", false);
}
examiner.prototype  = {
  observe: function(subject, topic, data) {
    // subject should be an nsURI
    if (!SpecialPowers.can_QI(subject))
      return;

    if (topic === "http-on-opening-request") {
      var asciiSpec = SpecialPowers.getPrivilegedProps(SpecialPowers.do_QueryInterface(subject, "nsIHttpChannel"), "URI.asciiSpec");
      if (asciiSpec !== reportURI) return;

      try {
        // Verify that the report was properly formatted.
        // We'll parse the report text as JSON and verify that the properties
        // have expected values.
        var reportText = "{}";
        var uploadStream = SpecialPowers.wrap(SpecialPowers.do_QueryInterface(subject, "nsIUploadChannel")).uploadStream;

        if (uploadStream) {
          // get the bytes from the request body
          var binstream = SpecialPowers.Cc["@mozilla.org/binaryinputstream;1"]
                                          .createInstance(SpecialPowers.Ci.nsIBinaryInputStream);
          binstream.setInputStream(uploadStream);

          var segments = [];
          for (var count = uploadStream.available(); count; count = uploadStream.available()) {
            var data = binstream.readBytes(count);
            segments.push(data);
          }

          var reportText = segments.join("");
          // rewind stream as we are supposed to - there will be an assertion later if we don't.
          SpecialPowers.do_QueryInterface(uploadStream, "nsISeekableStream").seek(SpecialPowers.Ci.nsISeekableStream.NS_SEEK_SET, 0);
        }
        try {
          var reportObj = JSON.parse(reportText);
        }
        catch (e) {
          ok(false, "Could not parse JSON (exception: " + e + ")");
        }
        // test for the proper values in the report object
        window.checkResults(reportObj);
      }
      catch (e) {
        ok(false, "Could not query report (exception: " + e + ")");
      }

      // finish up
      window.examiner.remove();
      SimpleTest.finish();
    }
  },

  // remove the listener
  remove: function() {
    SpecialPowers.removeObserver(this, "http-on-opening-request");
  }
}
window.examiner = new examiner();
SimpleTest.waitForExplicitFinish();

// load the resource which will generate a CSP violation report
// save this for last so that our listeners are registered.
var src = "file_csp_testserver.sjs";
// append the file that should be served
src += "?file=" + escape(testfile);
// append the CSP that should be used to serve the file
src += "&csp=" + escape(policy);
// appending a fragment so we can test that it's correctly stripped
// for document-uri and source-file.
src += "#foo";
document.getElementById("cspframe").src = src;

</script>
</pre>
</body>
</html>
